⚡ Bolt: Optimize slice allocations in nested append calls
Replaced nested `append` calls when copying and appending EOL sequences in `SendInput` for `protocol-rawsocket` and `protocol-serial` with single-allocation `make([]byte, 0, cap)` calls.

Cast `len(data)` to `int64` during pre-allocation to satisfy CodeQL integer overflow rules.

This avoids an unnecessary secondary allocation and copy. Tests have been executed to confirm behavior remains correct.

+++ plugins/protocol-rawsocket/session.go
@@ -137,7 +137,7 @@ func (s *Session) SendInput(ctx context.Context, data []byte) error {
 	case EOLModeLF:
 		if !bytes.HasSuffix(data, []byte("\n")) {
 			// Bolt: Optimization to avoid double allocation in nested append.
-			payload = make([]byte, 0, len(data)+1)
+			payload = make([]byte, 0, int64(len(data))+1)
 			payload = append(payload, data...)
 			payload = append(payload, '\n')
 		}
@@ -146,12 +146,12 @@ func (s *Session) SendInput(ctx context.Context, data []byte) error {
 			// If it ends with a bare LF, upgrade to CRLF; otherwise append CRLF.
 			if bytes.HasSuffix(data, []byte("\n")) {
 				// Bolt: Optimization to avoid double allocation in nested append.
-				payload = make([]byte, 0, len(data)+1)
+				payload = make([]byte, 0, int64(len(data))+1)
 				payload = append(payload, data[:len(data)-1]...)
 				payload = append(payload, '\r', '\n')
 			} else {
 				// Bolt: Optimization to avoid double allocation in nested append.
-				payload = make([]byte, 0, len(data)+2)
+				payload = make([]byte, 0, int64(len(data))+2)
 				payload = append(payload, data...)
 				payload = append(payload, '\r', '\n')
 			}
