+++ plugins/protocol-serial/session.go
@@ -140,7 +140,7 @@ func (s *Session) SendInput(ctx context.Context, data []byte) error {
 	case EOLModeLF:
 		if !bytes.HasSuffix(data, []byte("\n")) {
 			// Bolt: Optimization to avoid double allocation in nested append.
-			payload = make([]byte, 0, len(data)+1)
+			payload = make([]byte, 0, int64(len(data))+1)
 			payload = append(payload, data...)
 			payload = append(payload, '\n')
 		}
@@ -148,12 +148,12 @@ func (s *Session) SendInput(ctx context.Context, data []byte) error {
 		if !bytes.HasSuffix(data, []byte("\r")) {
 			if bytes.HasSuffix(data, []byte("\n")) {
 				// Bolt: Optimization to avoid double allocation in nested append.
-				payload = make([]byte, 0, len(data))
+				payload = make([]byte, 0, int64(len(data)))
 				payload = append(payload, data[:len(data)-1]...)
 				payload = append(payload, '\r')
 			} else {
 				// Bolt: Optimization to avoid double allocation in nested append.
-				payload = make([]byte, 0, len(data)+1)
+				payload = make([]byte, 0, int64(len(data))+1)
 				payload = append(payload, data...)
 				payload = append(payload, '\r')
 			}
@@ -162,12 +162,12 @@ func (s *Session) SendInput(ctx context.Context, data []byte) error {
 		if !bytes.HasSuffix(data, []byte("\r\n")) {
 			if bytes.HasSuffix(data, []byte("\n")) {
 				// Bolt: Optimization to avoid double allocation in nested append.
-				payload = make([]byte, 0, len(data)+1)
+				payload = make([]byte, 0, int64(len(data))+1)
 				payload = append(payload, data[:len(data)-1]...)
 				payload = append(payload, '\r', '\n')
 			} else {
 				// Bolt: Optimization to avoid double allocation in nested append.
-				payload = make([]byte, 0, len(data)+2)
+				payload = make([]byte, 0, int64(len(data))+2)
 				payload = append(payload, data...)
 				payload = append(payload, '\r', '\n')
 			}
